# Apply autofixes iteratively until code is stabilized