Resolves #868, #866, #660.