@JonathanPlasse 2022-11-22
How does it impact performance?