@charliermarsh 2022-11-22
Seems pretty fast! Certainly within the margin of error, possibly even faster because of some other changes in the PR?

The nice thing is that we only need to re-lint files that _changed_, and we don't actually have to write and read from disk every time we re-lint -- we can just keep the changed contents in memory.

Before:

```
∴ hyperfine --warmup 10 --runs 25 -i --cleanup "git reset --hard HEAD" "../../../target/release/ruff --no-cache --fix ."
Benchmark 1: ../../../target/release/ruff --no-cache --fix .
  Time (mean ± σ):     308.2 ms ±   7.5 ms    [User: 2475.2 ms, System: 92.0 ms]
  Range (min … max):   302.5 ms … 342.4 ms    25 runs
```

After:

```
∴ hyperfine --warmup 10 --runs 25 -i --cleanup "git reset --hard HEAD" "../../../target/release/ruff --no-cache --fix ."
Benchmark 1: ../../../target/release/ruff --no-cache --fix .
  Time (mean ± σ):     304.1 ms ±   2.7 ms    [User: 2449.3 ms, System: 87.2 ms]
  Range (min … max):   300.0 ms … 309.7 ms    25 runs
```